Hero removed: Invisible walls removed, hero object removed, player spawn removed, input script modified
