--- towerhero/Assets/Prefabs/InvisibleWall.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a043e6e033bb940eca0cfc2fb9680136
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
