--- towerhero/Assets/Prefabs/InvisibleWall10 1.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 30e7c6328734f4d54bdff4e668124f01
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
