--- towerhero/Assets/Prefabs/InvisibleWall10.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 869f8b623ddcc456cb84785f76f973ff
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
