--- towerhero/Assets/Prefabs/InvisibleWall11.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d5b233ed6c8ee4010901c355aec625de
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
