--- towerhero/Assets/Prefabs/InvisibleWall12.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2f146c4a3c6ea4eb9aab6253a4925d1e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
