--- towerhero/Assets/Prefabs/InvisibleWall13.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e583db3b9a1b344119cb7caaf219fefe
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
