--- towerhero/Assets/Prefabs/InvisibleWall14.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8ce5628b4b59f40e1b8d7391b37b8069
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
