--- towerhero/Assets/Prefabs/InvisibleWall15.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bba8c4044c0454d0d9a880ecc32a3315
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
