--- towerhero/Assets/Prefabs/InvisibleWall16.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 23c3aeaf920864bb4bcea5cce72684d8
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
