--- towerhero/Assets/Prefabs/InvisibleWall2.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6d5191e307959427cb01f02a8ac48321
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
