--- towerhero/Assets/Prefabs/InvisibleWall3.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0044e05e80ca41269635ed017351ca2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
