--- towerhero/Assets/Prefabs/InvisibleWall4.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 70b3f4bb4b50647f5ad7e686dd17edf5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
