--- towerhero/Assets/Prefabs/InvisibleWall5.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a5212d9a2b9b04da5acbfb984315c334
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
