--- towerhero/Assets/Prefabs/InvisibleWall6.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e7c04e4bcf57f447b96cc93cb7b1a5a3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
