--- towerhero/Assets/Prefabs/InvisibleWall7.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 15b9a73f02a164f2695dfff5885a6032
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
