--- towerhero/Assets/Prefabs/InvisibleWall8.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bcb94602448044431ba41ee608e41f22
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
