--- towerhero/Assets/Prefabs/InvisibleWall9.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e1d31faad1d2c49ea8fc1a83fe1410b4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
