--- towerhero/Assets/Prefabs/PlayerSpawn.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 49c306e09755d4b5bbf12cd0933dee5e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
