--- towerhero/Assets/Scripts/PlayerMovementScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a234438ec1f0a41bc8fdf3b89f147939
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
