--- towerhero/Assets/Scripts/PlayerSpawnScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cca40abefc7504c67ba95919b16a8bba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
